move the graph clustering functions to Toolbox.Util.Graph
